+++ internal/templates/blocks/utils.go
@@ -3,6 +3,8 @@ package blocks
 import (
 	"html/template"
 
+	"github.com/a-h/templ"
+	"github.com/kaugesaar/lucide-go"
 	"github.com/nathanhollows/Rapua/v6/helpers"
 )
 
@@ -12,4 +14,9 @@ func stringToMarkdown(s string) template.HTML {
 		return template.HTML(err.Error())
 	}
 	return template.HTML(md)
+}
+
+func icon(icon string, class string) templ.Component {
+	i := lucide.Icon(icon, map[string]any{"class": class})
+	return templ.Raw(i)
 }